+++ lib/device_init.h
@@ -1,0 +1,6 @@
+#ifndef DEVICE_INIT_H
+#define DEVICE_INIT_H
+
+int initialize_devices();
+
+#endif // DEVICE_INIT_H
